history_service/requirements.txt
grpcio==1.30.0
grpcio-tools
python-consul==1.1.0
